requirements.txt
# pip freeze
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
python-dotenv==1.2.1
requests==2.32.5
urllib3==2.6.3
black==26.1.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
flake8==7.3.0
idna==3.11
iniconfig==2.3.0
librt==0.7.8
mccabe==0.7.0
mypy==1.19.1
mypy_extensions==1.1.0
numpy==2.4.1
packaging==25.0
pandas==3.0.0
pathspec==1.0.3
platformdirs==4.5.1
pluggy==1.6.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.4.0
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
urllib3==2.6.3
